requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
